Revert "Fixed spaces"
This reverts commit fcc21eae95ea64ccbe54e82457d62d12d3a1f405.
Revert "Fixed indentation"

This reverts commit dbd163737a9cc1c48e5e55de08b8bf8356586335.
Revert "Changed from using floats to ints for doc terms & frequencies"

This reverts commit b1f004e19fed2cd9f9fd625eed24fb98405a6c05.

+++ gensim/models/ldamodel.py
@@ -698,9 +698,7 @@ class LdaModel(interfaces.TransformationABC, basemodel.BaseTopicModel):
             dirty = False
 
             reallen = 0
-            for chunk_no, chunk in enumerate(utils.grouper(
-                                                       corpus, chunksize, as_numpy=chunks_as_numpy,
-                                                       dtype=self.dtype)):
+            for chunk_no, chunk in enumerate(utils.grouper(corpus, chunksize, as_numpy=chunks_as_numpy)):
                 reallen += len(chunk)  # keep track of how many documents we've processed so far
 
                 if eval_every and ((reallen == lencorpus) or ((chunk_no + 1) % (eval_every * self.numworkers) == 0)):
